Add company information and FAQ components; implement custom media query hook
