+++ package.json
@@ -43,6 +43,7 @@
       "react-leaflet": "^5.0.0",
       "tailwind-merge": "^2.6.0",
       "tailwindcss-animate": "^1.0.7",
+      "usehooks-ts": "^3.1.0",
       "zod": "^3.24.1"
    },
    "devDependencies": {
