+++ src/lib/constants/FAQ.ts
@@ -1,0 +1,16 @@
+import { FAQ } from "@/types/ComponentTypes";
+
+export const faqs: FAQ[] = [
+    {
+      question: "What is your typical lead time for custom orders?",
+      answer: "Our custom orders typically take 6-8 weeks from design approval to delivery. Each piece is meticulously crafted to meet our exacting standards."
+    },
+    {
+      question: "Do you offer international shipping?",
+      answer: "Yes, we offer worldwide shipping through our trusted logistics partners. Shipping times and costs vary by destination."
+    },
+    {
+      question: "What is your return policy?",
+      answer: "We offer a 30-day return policy for unused items in original packaging. Custom orders are non-returnable unless defective."
+    }
+  ];
